style(header.component.ts): update header component to use Bootstrap navbar
The header component has been updated to use the Bootstrap navbar instead of a simple flexbox. This provides a more responsive and user-friendly navigation experience. The nav-brand has been added to the navbar to display the app name. The navbar-collapse and navbar-nav classes have been added to the navbar to enable the dropdown menu for the navigation links.